upload foto

+++ admin/DDokter.php
@@ -48,7 +48,7 @@ while($d = mysqli_fetch_array($data)){
       <td><?php echo $d['harga']; ?></td>
       <td><?php echo $d['gambar']; ?></td>
       <td><button type='button' class='btn btn-outline-primary'><a style='text-decoration: none; color:#465771;' href='update.php?id_dokter=<?=$d["id_dokter"]?>'>Edit</a></button></td>
-        <td><button type='button' class='btn btn-outline-danger'><a style='text-decoration: none;color:#ce3046;' href='hapus.php?id_dokter=<?=$d["id_dokter"]?>'>Hapus</a></button></td>
+        <td><button type='button' class='btn btn-outline-danger'><a style='text-decoration: none;color:#ce3046;' onclick="return confirm('Apakah anda yakin?');" href='hapus.php?id_dokter=<?=$d["id_dokter"]?>'>Hapus</a></button></td>
     </tr>
 <?php
 }
